[DEV] macem.swift: introduce macem with simple features
Macem is a command-line tool to export all certificates and their
private keys from keychain. This commit adds a new file `macem.swift`
which contains a command-line tool to export all certificates and their
private keys from keychain.

The tool exports the certificates to a .p12 file which can be shared
between hosts. The tool takes two arguments: `-P` for the password that
protects the resulting .p12 file and `-o` for the target file name. The
tool retrieves all identities whose label starts with "Apple " from
keychain and exports them to the specified file. If no identity is
found, the tool skips the export.